Add confirmation to proposal stepper
